Fix 2019/karns
A strange problem is that although it worked with macOS and linux for me
numerous times in the past it stopped working under both. I went to try
and figure it out but when disabling the optimiser for debug symbols the
problem vanished. Changing it back to -O3 and the problem reappeared
both with macOS and linux. Thus it appears that the entry should not
have -O level > 0. I'm not sure if this is the same issue the author
reported where it sometimes segfaults but I did not try and figure it
out either (in any event the crash disappeared when -O0 was specified so
it'd be harder to debug).

Added to .gitignore the file the author noted the program is called (as
the command he gives creates). To make it easier I just made it a
symlink to prog.

Fixed formatting in README.md.