Update all/README.md
Make markdown format. This fixes a number of formatting problems
including no links.

Update obfuscate definition to be more modern (yes I used an American
English dictionary for this! :-) ) and also added definition of
obfuscation. Naturally as a logophile I did indeed include the etymology
of both.

Update some text. I tried to keep the historically significance context
in the file but this is a work in progress for sure.

Changed last updated timestamp. As tempting as it was to write '4 April
...' I just used the output of date(1) in my time zone (which happens to
be the same as the original time zone, whatever that might or might not
mean).

Moved copyright stuff to bottom of file so that the casual reader can
see the contest specific information immediately.

I am aware of some improvements that can be made but later on - today or
another day. The changes in particular have to do with
building/compiling entries in modern systems. Grammar might also be
improvable but that's to be determined later on as well.