Fixes in all/summary.txt
Reword the lines in the form of 'There is no YYYY IOCCC contest' (which
was unintentionally redundant with IOCCC contest for last two words
contest contest :-) ) to be instead 'There was no IOCCC in YYYY'.

Added the years 2021 and 2022 to the file (in the form as above).

A question that remains is whether or not this file should be turned
into a markdown file. On the one hand a markdown file might be nice if
rendered to html for browsers but a text file is nice for text editor
(obviously vim :-) ).