Fix 1984/decot to not require -traditional-cpp
This was fun to fix. What prompted me to do so is that I noticed that
the patch from Yusuke Endoh actually doesn't work in fedora linux. It never
worked with clang as it required -traditional-cpp and also had
non-standard second and third args type to main (not a char ** but
implicit int). No more. This entry now works with both linux and macOS!

I was able to remove all the macros as they were problematic and not
needed. The code that used them was updated to be what was necessary.

Add to Makefile inclusion of stdio.h and math.h (printf, floor).
Some systems require -lm so that was also added to the Makefile.

Very little else was required to make it work.

The decot.alt.c is the version patched by Yusuke but fixed by me to
compile with fedora's gcc.

Removed this entry from bugs.md.