Fixes in 1985/lycklama/README.md + Makefile
Add alt target to Makefile. Won't work for most systems but the alt code
is there (particularly so people can see how it looked when run on
itself back then) so it might as well be possible to run make alt for
those who have an old compiler.

Format fixes in README.md.