Fix 1986/wall to no longer need -traditional-cpp
This was also a lot of fun to fix. It took a fair bit of tinkering but
it was a lot of fun to figure out and now it works like it does with
-traditional-cpp!

Removed this entry from the bugs.md file as it now works.

Also added to the bugs.md a minor clarification about another entry
that is in status INABIAF (about warnings).