Remove 1985/sicherman -traditional-cpp requirement
For some good old C-fashioned fun I recommend people to compare
sicherman.c to sicherman.alt.c!

To the best of my knowledge with this there are no more entries that
require -traditional-cpp! The configure scripts in 2004/hibachi and
2005/mynx don't seem to be required to compile the code.

It's possible that other entries do need it that I've not yet discovered
but as of now no known entries need it.

Removed entry from bugs.md.