Fix 2005/mynx directory wrt configure script
The configure script which was in the directory overwrote (naturally)
the Makefile. I moved the configure to source/. I also moved
configure.in to source/ and makefile.in to source/.

Updated Makefile so that running make will work. Probably no need for
configure but since it's part of the entry I kept it.