Fixes in 2019/mills
Format fixes and typo fixes in README.md. Some of the typo fixes were
very subtle (like a ' outside of `` where it should be in) but I
probably missed some as there are many terms I'm unfamiliar with. I'll
have to go through this file again but it's a good start.

Add to .gitignore file name referenced README.md.

I believe this completes the first pass of formatting and typos in 2019
entries (I haven't really looked at the year README.md files).