Remove unreferenced 1990/cmills.alt.c
As discussed on GitHub this is rarely needed for gets() to fgets()
change (which again is useful - but problematic either way - to prevent
warnings during compile, linking or execution of the program sometimes,
including with this entry, output interspersed with the program's
output).

There was no need to modify the Makefile as the alt target was already
not defined.

BTW: if the compiler compiles, the linker links and the user executes
does that make the compiler the jury, the linker the judge and user the
executioner? :-)