Add todo.md at top level
This is because there have been many things numerous times where I
should have noted what still needs to be done but did not. This allows
me to update the list too. There can be a second list of what's done or
else the list of things to do / check / fix can have sub-lists under
entries of those already done. For instance under the entry checking
formatting of all README.md files it might be good to have a list of
years already checked. I've done a few but many more need to be done.
Hopefully this list will be updated in time. This file can be deleted
prior to the repo being merged into the winners repo.