Fix description of bug in 1990/westley
Add to README.md that one thing (a crash without an arg) is at this time
not a bug but a feature.

Fix the statement that a negative number causes the entry to enter an
infinite loop. Instead any input that's not a number > 0 will cause the
program to enter an infinite loop.