Fix 1995/garry
Formatting fixes.

Remove address as there was a formatting issue. Besides it's probably
inaccurate by now. Kept the country.

There was a stray partial sentence in the judges' remarks that strayed
in (or to be more correct it was never moved properly) when I added the
try section. It's now been removed.

URL is dead so used the Internet Wayback Machine. Updated the URL in the
winners.csv file too. That might need to be done with others that I've
recently fixed but that will have to be checked on another time.