Update 1995/vanschnitz
Added spoiler.c from their uuencoded text. Noted in README.md that they
encourage readers to try and figure it out first but the idea (told to
me by Landon) is that the README.md should no longer (generally
speaking) hide details.

Format fixes in README.md.

Note that this is one of the entries noted in bugs.md in commit
704f7224c0520d822e9cc99d0edeef7f03dc6516 but I don't believe spoiler.c
is the same as the missing file. That might be me reading it wrong in
which case that can be removed from bugs.md but it appears to not be:
one is a spoiler and another is different functionality.