Fix 2011/goren for macOS (64-bit)
On a hunch I wondered if the ints in function S() needed to be long and
it did indeed solve the problem. Tested it under linux and then I
noticed the author stated that it does not work for 64-bits (except that
it does work in 64-bit linux). Therefore I test compiled it with linux
as 32-bit too and both work.

I also noticed in git diff that it had CRLF line endings so I removed
those as well.