Chocolate cake!
Updated the chocolate-cake.md file to stress important points (including
by making the ingredients - in the instructions parts - upper case to
make it stand out more). Some of the points are critical.

Added to 2020/ferguson2/.gitignore a file that is used to encrypt the
chocolate-cake.md (from 2020/ferguson1) for the Enigma entry (even
though anyone interested would use the 2020/ferguson1 plaintext
version). This is because it's easier to have than to look up the syntax
each time (of the recode program).