Typo fix in 2020/ferguson2/README.md
Extraneous ')' added by accident when adding a link yesterday.

I also, unrelated, clarified in a spot or two that one should use the
markdown files if reading on GitHub. This however is not consistent and
needs to be done throughout all files probably including other entries
(mostly README.md but there are certainly others).