Format and link fixes in 2020/ferguson2
To make this easier to see what it all is I'm doing this in a separate
pull request.

Some of the links were broken and I added some links as well. I tried to
only link to the same thing once (per section) but it's possible I
missed some.

Format fixes in some code blocks and probably other things too.

Clarify and reword some things as well.

And _this_ should be the final one for now unless I see a problem on
GitHub (I'll be checking the rendered output of the markdown files and
making any fixes if necessary).