Update and fixes in 1996/jonth
I note that this entry does not work in macOS even with XQuartz. The
invocation recommended caused a crash but I came up with an invocation
that made it go further along - but still also crash. I wrote in bugs.md
details about it as well as giving a workaround in macOS (involves
another system with X11 forwarding - no not xhost). There is some C
pre-processed code in bugs.md in case this helps.

The URL of the author is dead and the Internet Wayback Machine was of no
help so it was removed from winners.csv and winners.html as well. It was
added as a missing link in bugs.md. It was not in the earlier entry from
1993 (they won twice) so I did not put it there.

Format fixes in README.md.