Format + link fix 1998/banks/
His website is of course no longer at the university he went to.

I also linked (via the Internet Wayback Machine which was found in the
archive of his old website) an article about the flight simulator that
he linked. It seemed appropriate especially as this won 'Best of Show'.

The URL was fixed in winners.csv and winners.html; his email was updated
in winners.csv as well (didn't see it in winners.html).