Format and link fix in 2001/jason
His URL was dead but I found via the Internet Wayback Machine a blog. He
hasn't really updated it (one entry this year) since I think it was 2021
but it's better than what it went to before. He might be on GitHub with
the same handle but I haven't confirmed it so I won't link to it. The
URL was updated in winners.html and winners.csv.

Format fix README.md.