Fix 2004/anonymous/README.md
Format fixes are obvious but I also added links to the wonderful Tolkien
related things like tehta (no that's not a typo), tengwa(r)? and
everything else in the file.

Also added link to the author of the font it's based on as well as their
page on the font. Fixed name of author as well.

Fixed some typos.

Moved the 'BTW' in judges' remarks to below the 'p.s.'.

Fixed URL. I'm not sure on this one because the author wished to be
anonymous (a real shame!) but left a URL anyway. Had to use the Internet
Wayback Machine. Also fixed in winners.{html,csv}.