Format and URL fix 2004/gavare
Format fix README.md.

Add humour heading (as 'humor' as the author spelt it wrong :-) :-) )
instead of what was a '<humor>..</humor>'. This due this is not valid
html (guess why! :-) ) and so would not show up right in GitHub (or in
any other html document). Using a '### ...' seemed better than escaping
the <>s.

Dead link. Found new one via the Internet Wayback Machine.