2001/herrmann2 requires 32-bit to work completely
By instinct I suspected that the reason it did not work fully (some
invocations do work) is that it's being compiled as 64-bit. This proved
to be the case: 32-bit works in full. Thus make alt now compiles it as
64-bit and the default compiles it as 32-bit.

To prevent making the Makefile more complicated for the default (not the
alt) target we add -m32 and the alt does not do anything special.

Updated bugs.md as the bug status and other information about it has
obviously changed. Same with README.md.

Updated .gitignore.