Format fixes in 2004/schnitzi + .gitignore update
The reason the top level .gitignore is updated is because this entry
will create vi swap files and it seems that the glob belongs with the
others.

For instance after running this entry numerous times over numerous days:

    $ ls -1 .s*|wc -l
    52

I was not aware of them until it started getting to .su like .sux and
others.

I've noted this fact in the README.md file as well, explaining to those
who don't know how they're named, how to find them (or what they're
called).