Add alt URL to winners.csv
I put the field at the end of each line as it's quicker and easier to do
than to try and fit it in after the primary URL. Since it's a temporary
file it doesn't really matter either.

In this case I only know of two authors with two URLs (or at least two):
Daniel Vik and myself. Both were updated. My primary and alt URL were
also swapped. I'm not sure if I've done that in my entries but I'll
check that in time - still not sure exactly which order I want but it
seems like the right order might be the one in winners.csv. This is TBD
later.

Some lines still certainly have to be fixed and that's including that
there are still links not checked for deadness. Some links that have
already been updated are at the Internet Wayback Machine and it might be
possible that some of these have their own new domain (or domains) but
not discovered yet.

As for Daniel Vik and alt url: in a future pull request these will be
updated. I already have the files updated but I felt it belongs in a
different commit.