Add email to Daniel Vik in winners.csv
Added email address. This has not been done for the two entries in 2004.

There might have been an extraneous field removed but I'm not even sure;
too tired to tell now and too tired to remember (I won't be doing much
until I'm more awake so there's no problem there). All have to be
checked in time however and I probably will try and automate it at some
point as I don't really fancy going through 196 lines (1 for header +
196 authors), checking one by one and potentially getting lost between
commas.

Well okay I know my line is okay too so I can make it 195 but that's
still too many manual lines to sift through. But on the subject of lines
and for those who need a laugh, an old programming joke: what is the
difference between a programmer and a drug addict? The type of 'line'
when they say 'one more line and I'll quit!'. :-)