Fixes in 1984/decot
Updated fixes in README in a better order (not chronological but in a
way that shows how it shows that I fixed it to not require
-traditional-cpp and also fixed the type of args in main() that clang
enforces). The fix by Yusuke Endoh which only let it compile with gcc
(with -traditional-cpp) is below that. To the cynics: this is not
because I want my fix at the top but rather that my fixes make it more
portable which makes it easier to use. With my fix the other one is not
technically even needed but I want to give thanks anyway as he
contributed to it (albeit without knowing about it yet).

Fixed make alt rule so that it uses -traditional-cpp as it needs it.
This code will not work with clang because it doesn't support that
option and it has the type of main() args issue. That means it won't
work with macOS default compiler even gcc.