Add optional usleep() to 1985/lycklama
With a tip from Yusuke Endoh I noticed that if you slow down the call to
write() you see some fun output. Now to make it as close to possible as
the original (there already is alt code! - mostly for fun display) I
don't call usleep() unless the macro Z is not 0. This way someone can
do:

    make CDEFINE+="-DZ=1000" clobber all

for example but if they just do make clobber all it won't be slow at all
or it'll be as slow as it takes the CPU to run it.