Fixes in 2005/aidan and add alt code
The alt code is in the README.md. It felt like it should be alt code
based on what the author stated.

Fixed the perl test suite. The script hardcoded the name of the binary
but this was not 'aidan' as the author submitted it as 'sudoku-sf'. This
broke the script (though I did not test it before, only after).

Added to the list of files noted by the author those which were in the
tarball and changed the names of any incorrectly named files to be
correct.

The author stated that in the US sudoku is known not as sudoku but
rather Number Place. This however is not true so I removed it as it most
likely isn't confusing but still not right (I didn't believe the name,
hadn't seen it and did a search and found nothing of the sort).

Since the author is from the UK and since they used British English
spelling I changed 'period' to 'full stop' as it should be (the author
put a '.' in parentheses so it's not confusing to anyone who doesn't
know the term anyway).

Format fix README.md.

Typo fixes.

Update .gitignore due to alt code.