Format fix 2005/anon/README.md + add alt version
The alt version uses the same code but the make rule is different. It
disables the stty command. This is in case a shell does not support it.
It also allows to demonstrate a funny feature that the author documented
where after too many moves the program deliberately destroys its runtime
stack. This is because the program sometimes deliberately generates
unsolvable puzzles to hook you :-) and also to prevent you spending too
much time or from getting too frustrated (see also the part about
unsolvable puzzles being deliberately generated).