Format fix 2005/jetro/README.md
I also removed the information on how to install SDL and referred
(linked) to the faq which has the information instead. It seems better
to just have a note about it being needed and then referring to the faq
rather than lengthening each README.md file by explaining how to do it.
Perhaps this is not the right way to go about it but that can be decided
later.