More fixes in 2005/README.md + update todo.md
The todo.md change is related; see below.

This commit was needed because in commit
68a9d0adb5609f360a3cfe597399e9915c53d6ad I moved (the 'yank paste get
distracted and forget error' :-) ) comments on some of the losing
entries to below the comments on the winning entries (since that's what
most people are interested in after all) to the bottom but then forgot
(I have been awake since 0000 hours or a bit before!) to go through them
all.

Added links.

Possible spelling fixes.

I hope this finishes 2005. Note that the root README.md files for all
other years have NOT been fixed yet. I will do these in future, maybe
today or maybe not. I'm kind of focused on getting the entries done
first but I might do the next years like I did this year. This is TBD
but I've added it to the todo.md file.