Fix compilation and README of 2006/borsanyi
The Makefile needs -lpthread because not all systems implicitly link it
in. I tentatively noted this fix in the README.md but I'm not sure about
it as it's nothing special at all.

Format fix README.md. I note that I could not add <>s to the email
address as for some reason (in particular the dot in the email early on)
caused rdiscount (or I think - vim syntax certainly had a problem).