Format and URL fix 2006/sykes1 + add missing pdf
The pdf file is available on the Internet Wayback Machine. It felt like
it might be nice to have it stored in the repo instead since the link is
technically dead.

Some links (besides the pdf one) were dead. One I removed as I'm not
sure it has any purpose to have the Internet Wayback Machine archive.
Others were fixed (for example a YouTube video was no longer there so I
found an alternative).

Format fix README.md.