Format fix + add try script to 2006/sykes2
The script runs the 'try' section but makes sure to redirect both stderr
and stdout to /dev/null as without it it ruins the show.

Format fixed README.md.

The author suggested to having a table of operator precedence open when
looking at the code so I pointed out operator(7).